requirements.txt
customtkinter
fpdf
openai
python-dotenv
matplotlib